--- merp_outgoing_routing/__init__.py
@@ -1,4 +1,0 @@
-# Copyright 2019 VentorTech OU
-# Part of Ventor modules. See LICENSE file for full copyright and licensing details.
-
-from . import models
