+++ outgoing_routing/__init__.py
@@ -1,0 +1,5 @@
+# Copyright 2020 VentorTech OU
+# License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl-3.0).
+
+from . import models
+from . import tests
