+++ outgoing_routing/data/product_removal.xml
