+++ outgoing_routing/models/res_company.py
@@ -1,5 +1,5 @@
-﻿# Copyright 2019 VentorTech OU
-# Part of Ventor modules. See LICENSE file for full copyright and licensing details.
+﻿# Copyright 2020 VentorTech OU
+# License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl-3.0).
 
 from odoo import models, fields
 
@@ -25,7 +25,7 @@ class Company(models.Model):
 
     stock_reservation_strategy = fields.Selection(
         [
-            ('base', 'Routing Strategy'),
+            ('base', 'By Picking Strategy'),
             ('quantity', 'By Quantity'),
             ('none', 'Default'),
         ],
