+++ outgoing_routing/models/res_config.py
@@ -1,5 +1,5 @@
-﻿# Copyright 2019 VentorTech OU
-# Part of Ventor modules. See LICENSE file for full copyright and licensing details.
+﻿# Copyright 2020 VentorTech OU
+# License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl-3.0).
 
 from odoo import models, fields
 
@@ -8,31 +8,16 @@ class StockConfigSettings(models.TransientModel):
     _inherit = 'res.config.settings'
 
     outgoing_routing_strategy = fields.Selection(
-        [
-            # path should be valid for both stock pickings and quants
-            ('location_id.removal_prio', 'Location removal priority'),
-            ('location_id.name', 'Location name'),
-            ('product_id.name', 'Product name'),
-        ],
         string='Picking Strategy', default='location_id.name',
         related='company_id.outgoing_routing_strategy',
         readonly=False)
 
     outgoing_routing_order = fields.Selection(
-        [
-            ('0', 'Ascending (A-Z)'),
-            ('1', 'Descending (Z-A)'),
-        ],
         string='Picking Order', default='0',
         related='company_id.outgoing_routing_order',
         readonly=False)
 
     stock_reservation_strategy = fields.Selection(
-        [
-            ('base', 'Routing Strategy'),
-            ('quantity', 'By Quantity'),
-            ('none', 'Default'),
-        ],
         string='Reservation Strategy', default='base',
         related='company_id.stock_reservation_strategy',
         readonly=False)
