+++ outgoing_routing/models/stock_location.py
@@ -1,5 +1,5 @@
-﻿# Copyright 2019 VentorTech OU
-# Part of Ventor modules. See LICENSE file for full copyright and licensing details.
+﻿# Copyright 2020 VentorTech OU
+# License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl-3.0).
 
 from odoo import models, fields, api
 
@@ -40,7 +40,7 @@ class StockLocation(models.Model):
         remaining_locations = self - processed
         max_seq = len(res)
         for remaining in remaining_locations:
-            remaining.strategy_sequence =  max_seq
+            remaining.strategy_sequence = max_seq
 
     @api.onchange('location_id')
     def _onchange_parent_location(self):
