+++ outgoing_routing/tests/__init__.py
@@ -1,5 +1,5 @@
-# Copyright 2019 VentorTech OU
-# Part of Ventor modules. See LICENSE file for full copyright and licensing details.
+# Copyright 2020 VentorTech OU
+# License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl-3.0).
 
 from . import test_merp_outgoing_routing
 from . import test_merp_quants_location_routing
