+++ outgoing_routing/views/picking.xml
@@ -7,7 +7,7 @@
         <field name="inherit_id" ref="stock.view_picking_form"/>
         <field name="arch" type="xml">
             <xpath expr="//field[@name='move_line_ids_without_package']/.." position="after">
-                <page string="To Process" name="operations_to_pick" groups="merp_custom_access_rights.merp_debug">
+                <page string="To Process" name="operations_to_pick" groups="ventor_base.merp_debug">
                     <field name="strategy_order_r" readonly="1" />
                     <field name="operations_to_pick" readonly="1">
                         <tree>
