+++ outgoing_routing/views/report_stockpicking.xml
