+++ outgoing_routing/views/stock.xml
